requirements.txt
##################################
#           PRODUCTION           #
##################################

##### Frameworks #####

Flask

##### Debugging, Logging, and Documentation #####

##### System #####

setuptools
python-dotenv
blinker

##### Storage #####

PyGithub
pygit2
pillow
redis

##### Frontend #####

jsmin

##### Network #####

requests
gunicorn

##### Security #####

Authlib

##### Async #####
rq

##### AI #####

openai >= 1.42

##### Parsing #####
dateparser
python-slugify
